IsDesktopRestaurantsHome
made IsDesktopRestaurantsHome component that both chef of the week and Home use to make them leaner.
