+++ src/components/homepage/ChefOfTheweek/ChefOfTheweek.scss
@@ -70,7 +70,7 @@
   }
 
   
-  .RestaurantsHome {
+  .RestaurantsHomeDesktop {
     align-items: flex-start;
     margin-top: 45px;
     margin-left: 100px;
