+++ src/components/homepage/RestaurantsHomeDesktop/RestaurantsHomeDesktop.scss
@@ -1,6 +1,6 @@
 @import "/src/assets/styles/_mixins.scss";
 
-.RestaurantsHome {
+.RestaurantsHomeDesktop {
     max-width: 1440px;
     margin: 0 auto;
     margin-top: 50px;
