+++ src/pages/Home.tsx
@@ -4,17 +4,14 @@ import { mockDishes, mockRestaurants, mockChefs } from "../mockData"
 import Hero from '../components/homepage/Hero/Hero';
 import DishIconsList from '../components/homepage/DishIconsList/DishIconsList';
 import AboutUs from '../components/homepage/AboutUs/AboutUs';
-import RestaurantsSwiper from '../components/homepage/ResaurantsSwiper/RestaurantsSwiper';
 import ChefOfTheWeek from '../components/homepage/ChefOfTheweek/ChefOfTheweek';
 import { POPULAR_RESTAURANTS_IN_EPICURE } from '../utils/constants';
-import RestaurantsHome from '../components/homepage/RestaurantsHome/RestaurantsHome';
 import { RoutePaths } from '../utils/enum';
-import useIsDesktop from '../components/Common/useIsDesktop';
+import IsDesktopRestaurantsHome from '../components/homepage/IsDesktopRestaurantsHome';
 
 
 const Home: React.FC = () => {
 
-  const IsDesktopScreen = useIsDesktop();
 
   const yossiShitrit = mockChefs.find(chef => chef.name === "Yossi Shitrit");
 
@@ -24,21 +21,15 @@ const Home: React.FC = () => {
   return (
     <div className="home">
       <Hero/>
-      {IsDesktopScreen ? (
-        <RestaurantsHome
-          restaurants={mockRestaurants} 
-          title={POPULAR_RESTAURANTS_IN_EPICURE} 
-          allLink = {RoutePaths.RESTAURANTS}
-          showChef
-          isStars
-        />
-      ) : (
-        <RestaurantsSwiper 
-          restaurants={mockRestaurants} 
-          title={POPULAR_RESTAURANTS_IN_EPICURE} 
-          showChef
-        />
-      )}
+
+      <IsDesktopRestaurantsHome 
+        restaurants={mockRestaurants} 
+        title1={POPULAR_RESTAURANTS_IN_EPICURE} 
+        allLink = {RoutePaths.RESTAURANTS}
+        showChef
+        isStars
+      />
+  
       <DishSwiper dishes={signatureDishes} />
       <DishIconsList/>
       {yossiShitrit && <ChefOfTheWeek chef={yossiShitrit} />}
@@ -48,3 +39,4 @@ const Home: React.FC = () => {
 };
 
 export default Home;
+
